requirements.txt
asgiref==3.9.1
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.0
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
sqlparse==0.5.3

# Analytics and Async Processing Dependencies
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Data Science and Analytics - Updated for Python 3.13 compatibility
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
statsmodels>=0.14.1
scipy>=1.11.0

# Additional dependencies for async processing
asyncio-mqtt==0.16.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# PDF generation and visualization
reportlab==4.0.7
matplotlib>=3.8.0

# Two-Factor Authentication
pyotp==2.9.0
qrcode==7.4.2
